Refactored to use 32 bit index in field -- because 64 was CRAZY. Refactored production code.

+++ src/bit_field.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-void bit_field_set_bit(uint64_t *bits, uint64_t bit_to_set);
-int bit_field_is_set(const uint64_t *bits, uint64_t bit_to_check);
+void bit_field_set_bit(uint64_t *bits, uint32_t bit_to_set);
+int bit_field_is_set(const uint64_t *bits, uint32_t bit_to_check);
 
 #endif // BIT_FIELD_H
